+++ client/src/features/timeline/TimelineContainer.tsx
@@ -60,6 +60,19 @@ const TimelineContainer: React.FC<{
 
   const { data, isSuccess, isError } = useGetTimelineQuery(links)
 
+  // Checks if the element is in fullscreen
+  useEffect(() => {
+    const handleFullscreenChange = (): void => {
+      setIsFullscreen(document.fullscreenElement === timelineRef.current)
+    }
+
+    document.addEventListener('fullscreenchange', handleFullscreenChange)
+
+    return () => {
+      document.removeEventListener('fullscreenchange', handleFullscreenChange)
+    }
+  }, [])
+
   const setFullscreen = (): void => {
     setIsFullscreen(!isFullscreen)
     const elem = timelineRef.current